fix: fixed the errors, just had to separate enum to different class. DUMB
