+++ game/src/game/components/Pickups/PickupType.h
@@ -1,0 +1,20 @@
+#pragma once 
+
+enum class PowerupPickupType
+{
+    kDefaultPowerup = 0,
+    kDisableHandling,
+    kEveryoneSlower,
+    kIncreaseAimBox,
+    kKillAbilities
+};
+
+enum class AmmoPickupType
+{
+    kDefaultAmmo = 0,
+    kBuckshot,
+    kDoubleDamage,
+    kExploadingBullet,
+    kIncreaseFireRate,
+    kVampireBullet
+};
